Sep 12, 2024, 10:38 PM

+++ Lab2/MultimodDrivers/src/multimod_i2c.c
@@ -90,7 +90,7 @@ uint8_t I2C_ReadSingle(uint32_t mod, uint8_t addr)
 
     // Trigger I2C module receive
     I2CMasterControl(mod, I2C_MASTER_CMD_SINGLE_RECEIVE);
-    uint32_t data = I2CMasterDataGet(mod);
+    uint32_t data = (uint8_t)I2CMasterDataGet(mod);
 
     // Wait until I2C module is no longer busy
     while (I2CMasterBusy(mod))
